bloons_already_dmged changed, Boomerang is now subclass Projectile, buttons keys added
if two projectiles (dmg ==1 ) one of top on another were hitting a mob, the first one would pop the 1st layer then stop because it detects that it has already hit that mob (his parent), and the other one would pop another layer and that's it based on that same principle, but the next frame, projectile 1 doesn't know it already hit the mob grand grand child (he knows he hurted the mob and it's child but that's all). Therefore projectile 1 would pop another layer of the mob, but then, projectile 2 doesn't have that new child listed as already dmged.

To fix that, I created a list of all the projectiles that hurted the bloon that is hereditary. Then before popping a bloon, we check if we didn't already hurt him during last frames (that info is stored in the instance of the projectile) and if we didn't hurt him on that very frame (that info is stored on the mob instance).
Then, for every mob update, we add the mob in the already dmged mob list of all the projectiles that hurted him during that frame.